+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.3
+
+- **BREAKING**: Remove `appendExecutablePath` in `CommandInput`.
+- **BREAKING**: `CommandInput.arguments` now is parameterized type.
+- **FEATURE**: Public the `executablePath` in `CommandInput`.
+- **FEATURE**: Support the add new argument in `CommandInput`.
+
 ## 0.0.2
 
 - **[FIX]**: Fix the group command display sys message.